+++ test/test_helper.rb
@@ -32,12 +32,12 @@ end
 
 class Context
   attr_writer :virtual_path
-  
+
   def initialize
     @_assigns = {}
     @virtual_path = nil
   end
-  
+
   def assigns
     @_assigns
   end
@@ -47,4 +47,11 @@ class Context
   end
 end
 
-User = Struct.new(:id, :name, :sex)
+class User
+  attr_accessor :id, :name, :sex
+  def initialize(id=nil, name=nil, sex=nil)
+    @id = id
+    @name = name
+    @sex = sex
+  end
+end